chore: 폴더구조에 맞게 변경
