+++ src/route/Setting.tsx
@@ -1,6 +1,6 @@
-import Setting from "@app/setting/index/Setting";
+import Setting from "@app/setting/page";
 import { PATH } from "@route/path.ts";
-import ProfileEdit from "@app/setting/edit-profile/ProfileEdit";
+import ProfileEdit from "@app/setting/edit-profile/page";
 
 const SETTING_ROUTES = [
   { path: PATH.SETTING.ROOT, element: <Setting /> },
